# labs/K3440/Габов Михаил/lab1/backend/services/recipes/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RecipeStepResponseDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"stepNumber": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"stepNumber",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IngredientResponseDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecipeIngredientResponseDto": {
				"properties": {
					"amount": {
						"type": "string"
					},
					"ingredient": {
						"$ref": "#/components/schemas/IngredientResponseDto"
					}
				},
				"required": [
					"amount",
					"ingredient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecipeDetailResponseDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"difficulty": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"properties": {
							"username": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"username",
							"id"
						],
						"type": "object"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/RecipeStepResponseDto"
						},
						"type": "array"
					},
					"recipeIngredients": {
						"items": {
							"$ref": "#/components/schemas/RecipeIngredientResponseDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"difficulty",
					"createdAt",
					"user",
					"steps",
					"recipeIngredients"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecipeIngredientCreateDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecipeStepCreateDto": {
				"properties": {
					"stepNumber": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					}
				},
				"required": [
					"stepNumber",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecipeCreateDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"difficulty": {
						"type": "string"
					},
					"ingredients": {
						"items": {
							"$ref": "#/components/schemas/RecipeIngredientCreateDto"
						},
						"type": "array"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/RecipeStepCreateDto"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"difficulty"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecipeListResponseDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"difficulty": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"properties": {
							"username": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"username",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"difficulty",
					"createdAt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecipeUpdateDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"difficulty": {
						"type": "string"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/RecipeStepCreateDto"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IngredientDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IngredientCreateDto": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IngredientUpdateDto": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Recipes API",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
		"/recipes": {
			"post": {
				"operationId": "CreateRecipe",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeDetailResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecipeCreateDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetRecipes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RecipeListResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ingredient",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/recipes/{recipeId}": {
			"get": {
				"operationId": "GetRecipeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeDetailResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeDetailResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecipeUpdateDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ingredients": {
			"post": {
				"operationId": "CreateIngredient",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IngredientDto"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredients"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IngredientCreateDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllIngredients",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IngredientDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredients"
				],
				"security": [],
				"parameters": []
			}
		},
		"/ingredients/{ingredientId}": {
			"get": {
				"operationId": "GetIngredientById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IngredientDto"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "ingredientId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateIngredient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IngredientDto"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "ingredientId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IngredientUpdateDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteIngredient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "ingredientId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}