# labs/K3440/Габов Михаил/lab1/backend/services/recipes/src/dtos/recipe.dto.ts
export class RecipeStepCreateDto {
    stepNumber!: number;
    description!: string;
    imageUrl?: string;
}

export class RecipeIngredientCreateDto {
    id!: number; 
    amount!: string;
}

export class RecipeCreateDto {
    title!: string;
    description!: string;
    difficulty!: string;
    ingredients?: RecipeIngredientCreateDto[];
    steps?: RecipeStepCreateDto[];
}

export class RecipeUpdateDto {
    title?: string;
    description?: string;
    difficulty?: string;
    steps?: RecipeStepCreateDto[];
    ingredients?: RecipeIngredientCreateDto[];
}


export class IngredientResponseDto {
    id!: number;
    name!: string;
}

export class RecipeIngredientResponseDto {
    amount!: string;
    ingredient!: IngredientResponseDto;
}

export class RecipeStepResponseDto {
    id!: number;
    stepNumber!: number;
    description!: string;
    imageUrl?: string;
}

export class RecipeDetailResponseDto {
    id!: number;
    title!: string;
    description!: string;
    difficulty!: string;
    createdAt!: Date;
    user!: { id: number; username: string };
    steps!: RecipeStepResponseDto[];
    recipeIngredients!: RecipeIngredientResponseDto[];
}

export class RecipeListResponseDto {
    id!: number;
    title!: string;
    description!: string;
    difficulty!: string;
    createdAt!: Date;
    user!: { id: number; username: string };
}